Refactor out common implementation between different seed types
Also cleaned up require's on the way, no module requires underscore anymore.
