+++ lib/directorySeed.js
@@ -2,13 +2,9 @@
 
 var fs = require('fs'),
   path = require('path'),
-  mongodb = require("mongodb"),
-  Server = mongodb.Server,
-  Db = mongodb.Db,
   json2Mongo = require('json2mongo'),
-  MongoClient = mongodb.MongoClient,
   async = require('async'),
-  _ = require('underscore');
+  baseSeed = require('./baseSeed');
 
 module.exports = function (host, port, db, seedLocation, callback) {
   fs.readdir(seedLocation, function (err, files) {
@@ -16,7 +12,7 @@ module.exports = function (host, port, db, seedLocation, callback) {
       return callback(err);
     }
 
-
+    var combinedContent = {}
     async.each(files, function (file, done) {
       if (path.extname(file) != '.json') done(new Error("Invla"));
 
@@ -24,22 +20,13 @@ module.exports = function (host, port, db, seedLocation, callback) {
       var content = fs.readFileSync(seedLocation + '/' + file);
       if (!content) return done();
 
-      content = JSON.parse(content);
-
-      MongoClient.connect('mongodb://' + host + ':' + port + '/' + db, function (err, db) {
-        db.createCollection(collectionName, function (err, col) {
-          db.collection(collectionName, function (err, collection) {
-
-            content = json2Mongo(content);
-
-            collection.insertMany(content, {w: 1}, function (err) {
-              done(err);
-            });
-          });
-
-        });
-
-      });
-    }, callback);
+      combinedContent[collectionName] = json2Mongo(JSON.parse(content));
+      done();
+    }, function(err) {
+      if (err) {
+        return callback(err);
+      }
+      baseSeed(host, port, db, combinedContent, callback);
+    });
   });
 };
