+++ lib/fileSeed.js
@@ -2,11 +2,8 @@
 
 var fs = require('fs'),
   path = require('path'),
-  mongodb = require("mongodb"),
-  async = require('async'),
   json2Mongo = require('json2mongo'),
-  MongoClient = mongodb.MongoClient,
-  _ = require('underscore');
+  baseSeed = require('./baseSeed');
 
 module.exports = function (host, port, db, seedFile, callback) {
 
@@ -22,18 +19,10 @@ module.exports = function (host, port, db, seedFile, callback) {
   content = JSON.parse(content);
   var isExtended = content.dataFormat === "EXTENDED_JSON";
   delete content.dataFormat;
-  async.each(Object.keys(content), function (collectionName, done) {
-    MongoClient.connect('mongodb://' + host + ':' + port + '/' + db, function (err, db) {
-      db.createCollection(collectionName, function (err, col) {
-        db.collection(collectionName, function (err, collection) {
 
-          var data = isExtended ? json2Mongo(content[collectionName]) : content[collectionName];
-
-          collection.insertMany(data, {w: 1}, function (err) {
-            done(err);
-          });
-        });
-      });
-    });
-  }, callback);
+  Object.keys(content).forEach(function (key) {
+    var value = content[key];
+    content[key] = isExtended ? json2Mongo(value) : value;
+  });
+  baseSeed(host, port, db, content, callback);
 };
